Move dev scripts and tests. Bring them up-to-date.
