+++ scripts/check_generated_games.py
